+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.1.3 (unreleased)
 
 - Added front-end pages.
+- Added cover-art scraping.